requirements.txt
# coloredlogs==15.0.1
# contourpy==1.3.0
# cycler==0.12.1
# filelock==3.17.0
# flatbuffers==23.3.3
# fonttools==4.55.8
# fsspec==2025.2.0
# hpack==4.0.0
# humanfriendly==10.0
# importlib_resources==6.5.2
# Jinja2==3.1.5
# kiwisolver==1.4.7
# lsq-ellipse==2.2.1
# MarkupSafe==3.0.2
# matplotlib==3.9.4
# mpmath==1.3.0
# munkres==1.1.4
# networkx==3.2.1
# numpy==2.0.2
# nvidia-cublas-cu12==12.1.3.1
# nvidia-cuda-cupti-cu12==12.1.105
# nvidia-cuda-nvrtc-cu12==12.1.105
# nvidia-cuda-runtime-cu12==12.1.105
# nvidia-cudnn-cu12==8.9.2.26
# nvidia-cufft-cu12==11.0.2.54
# nvidia-curand-cu12==10.3.2.106
# nvidia-cusolver-cu12==11.4.5.107
# nvidia-cusparse-cu12==12.1.0.106
# nvidia-nccl-cu12==2.20.5
# nvidia-nvjitlink-cu12==12.8.61
# nvidia-nvtx-cu12==12.1.105
# opencv-python==4.10.0.84
# packaging==24.2
# pillow==11.1.0
# pyparsing==3.2.1
# PyQt5==5.15.9
# PyQt5-Qt5==5.15.16
# PyQt5-sip==12.12.2
# python-dateutil==2.9.0.post0
# PyYAML==6.0.2
# scipy==1.13.1
# shapely==2.0.7
# six==1.17.0
# sympy==1.13.3
# torch==2.3.1
# torchaudio==2.3.1
# torchvision==0.18.1
# tqdm==4.67.1
# triton==2.3.1
# typing_extensions==4.12.2
# zipp==3.21.0
# zstandard==0.22.0


# torch==2.3.1
# torchaudio==2.3.1
# torchvision==0.18.1
# opencv-python==4.10.0.84
# scipy==1.13.1
# lsq-ellipse==2.2.1
# numpy==1.24.2
# pyyaml==6.0.2
# matplotlib==3.8.4
# tqdm==4.67.1
# pillow==11.1.0



colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.3.0
cycler==0.12.1
filelock==3.17.0
flatbuffers==23.3.3
fonttools==4.55.8
fsspec==2025.2.0
hpack==4.0.0
humanfriendly==10.0
importlib_resources==6.5.2
intel-openmp==2021.4.0
Jinja2==3.1.5
kiwisolver==1.4.7
lsq-ellipse==2.2.1
MarkupSafe==3.0.2
matplotlib==3.9.4
mkl==2021.4.0
mpmath==1.3.0
munkres==1.1.4
networkx==3.2.1
numpy==1.24.3
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.7.1.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nvjitlink-cu12==12.8.61
nvidia-nvtx-cu12==12.1.105
opencv-python==4.10.0.84
packaging==24.2
pillow==11.1.0
pyparsing==3.2.1
PyQt5==5.15.9
PyQt5-Qt5==5.15.2
PyQt5-sip==12.12.2
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
PyYAML==6.0.2
scipy==1.13.1
shapely==2.0.7
six==1.17.0
sympy==1.13.1
tbb==2021.13.1
torch==2.5.1+cu121
torchaudio==2.5.1+cu121
torchvision==0.20.1+cu121
tqdm==4.67.1
typing_extensions==4.12.2
zipp==3.21.0
zstandard==0.22.0
pykalman==0.10.1




# Dependencies pour l'API
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0